pyproject.toml
[tool.poetry]
name = "document_manager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
Django = "^3.2.7"
django-bridger = "0.11.24"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
